requirements.txt
Flask==2.0.1
pandas==1.3.3
python-docx==0.8.11
PyPDF2==2.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
nltk==3.6.3
reportlab==3.6.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1 